Rename to singular BoxParser for consistency with other types like ColorParser
