+++ libraries/extractor/src/main/java/androidx/media3/extractor/mp4/FragmentedMp4Extractor.java
@@ -19,7 +19,7 @@ import static androidx.media3.common.util.Assertions.checkNotNull;
 import static androidx.media3.common.util.Assertions.checkState;
 import static androidx.media3.common.util.Util.castNonNull;
 import static androidx.media3.common.util.Util.nullSafeArrayCopy;
-import static androidx.media3.extractor.mp4.BoxParsers.parseTraks;
+import static androidx.media3.extractor.mp4.BoxParser.parseTraks;
 import static java.lang.Math.max;
 import static java.lang.annotation.ElementType.TYPE_USE;
 
@@ -778,7 +778,7 @@ public class FragmentedMp4Extractor implements Extractor {
     }
     atom.setPosition(Mp4Box.HEADER_SIZE);
     int fullAtom = atom.readInt();
-    int version = BoxParsers.parseFullBoxVersion(fullAtom);
+    int version = BoxParser.parseFullBoxVersion(fullAtom);
     String schemeIdUri;
     String value;
     long timescale;
@@ -884,7 +884,7 @@ public class FragmentedMp4Extractor implements Extractor {
   private static long parseMehd(ParsableByteArray mehd) {
     mehd.setPosition(Mp4Box.HEADER_SIZE);
     int fullAtom = mehd.readInt();
-    int version = BoxParsers.parseFullBoxVersion(fullAtom);
+    int version = BoxParser.parseFullBoxVersion(fullAtom);
     return version == 0 ? mehd.readUnsignedInt() : mehd.readUnsignedLongToLong();
   }
 
@@ -1006,7 +1006,7 @@ public class FragmentedMp4Extractor implements Extractor {
     int vectorSize = encryptionBox.perSampleIvSize;
     saiz.setPosition(Mp4Box.HEADER_SIZE);
     int fullAtom = saiz.readInt();
-    int flags = BoxParsers.parseFullBoxFlags(fullAtom);
+    int flags = BoxParser.parseFullBoxFlags(fullAtom);
     if ((flags & 0x01) == 1) {
       saiz.skipBytes(8);
     }
@@ -1050,7 +1050,7 @@ public class FragmentedMp4Extractor implements Extractor {
   private static void parseSaio(ParsableByteArray saio, TrackFragment out) throws ParserException {
     saio.setPosition(Mp4Box.HEADER_SIZE);
     int fullAtom = saio.readInt();
-    int flags = BoxParsers.parseFullBoxFlags(fullAtom);
+    int flags = BoxParser.parseFullBoxFlags(fullAtom);
     if ((flags & 0x01) == 1) {
       saio.skipBytes(8);
     }
@@ -1062,7 +1062,7 @@ public class FragmentedMp4Extractor implements Extractor {
           "Unexpected saio entry count: " + entryCount, /* cause= */ null);
     }
 
-    int version = BoxParsers.parseFullBoxVersion(fullAtom);
+    int version = BoxParser.parseFullBoxVersion(fullAtom);
     out.auxiliaryDataPosition +=
         version == 0 ? saio.readUnsignedInt() : saio.readUnsignedLongToLong();
   }
@@ -1084,7 +1084,7 @@ public class FragmentedMp4Extractor implements Extractor {
       ParsableByteArray tfhd, SparseArray<TrackBundle> trackBundles, boolean haveSideloadedTrack) {
     tfhd.setPosition(Mp4Box.HEADER_SIZE);
     int fullAtom = tfhd.readInt();
-    int atomFlags = BoxParsers.parseFullBoxFlags(fullAtom);
+    int atomFlags = BoxParser.parseFullBoxFlags(fullAtom);
     int trackId = tfhd.readInt();
     @Nullable
     TrackBundle trackBundle =
@@ -1133,7 +1133,7 @@ public class FragmentedMp4Extractor implements Extractor {
   private static long parseTfdt(ParsableByteArray tfdt) {
     tfdt.setPosition(Mp4Box.HEADER_SIZE);
     int fullAtom = tfdt.readInt();
-    int version = BoxParsers.parseFullBoxVersion(fullAtom);
+    int version = BoxParser.parseFullBoxVersion(fullAtom);
     return version == 1 ? tfdt.readUnsignedLongToLong() : tfdt.readUnsignedInt();
   }
 
@@ -1176,7 +1176,7 @@ public class FragmentedMp4Extractor implements Extractor {
       throws ParserException {
     trun.setPosition(Mp4Box.HEADER_SIZE);
     int fullAtom = trun.readInt();
-    int atomFlags = BoxParsers.parseFullBoxFlags(fullAtom);
+    int atomFlags = BoxParser.parseFullBoxFlags(fullAtom);
 
     Track track = trackBundle.moovSampleTable.track;
     TrackFragment fragment = trackBundle.fragment;
@@ -1287,7 +1287,7 @@ public class FragmentedMp4Extractor implements Extractor {
       throws ParserException {
     senc.setPosition(Mp4Box.HEADER_SIZE + offset);
     int fullAtom = senc.readInt();
-    int flags = BoxParsers.parseFullBoxFlags(fullAtom);
+    int flags = BoxParser.parseFullBoxFlags(fullAtom);
 
     if ((flags & 0x01 /* override_track_encryption_box_parameters */) != 0) {
       // TODO: Implement this.
@@ -1340,7 +1340,7 @@ public class FragmentedMp4Extractor implements Extractor {
     }
 
     sbgp.setPosition(Mp4Box.HEADER_SIZE);
-    int sbgpVersion = BoxParsers.parseFullBoxVersion(sbgp.readInt());
+    int sbgpVersion = BoxParser.parseFullBoxVersion(sbgp.readInt());
     sbgp.skipBytes(4); // grouping_type == seig.
     if (sbgpVersion == 1) {
       sbgp.skipBytes(4); // grouping_type_parameter.
@@ -1351,7 +1351,7 @@ public class FragmentedMp4Extractor implements Extractor {
     }
 
     sgpd.setPosition(Mp4Box.HEADER_SIZE);
-    int sgpdVersion = BoxParsers.parseFullBoxVersion(sgpd.readInt());
+    int sgpdVersion = BoxParser.parseFullBoxVersion(sgpd.readInt());
     sgpd.skipBytes(4); // grouping_type == seig.
     if (sgpdVersion == 1) {
       if (sgpd.readUnsignedInt() == 0) {
@@ -1408,7 +1408,7 @@ public class FragmentedMp4Extractor implements Extractor {
       throws ParserException {
     atom.setPosition(Mp4Box.HEADER_SIZE);
     int fullAtom = atom.readInt();
-    int version = BoxParsers.parseFullBoxVersion(fullAtom);
+    int version = BoxParser.parseFullBoxVersion(fullAtom);
 
     atom.skipBytes(4);
     long timescale = atom.readUnsignedInt();
